requirements.txt
bleach==1.5.0
certifi==2016.2.28
cycler==0.10.0
deap==1.0.2
html5lib==0.9999999
Keras==2.0.8
Markdown==2.6.9
matplotlib==2.1.0
mss==3.0.1
numpy==1.13.3
pip==9.0.1
protobuf==3.4.0
PyBrain==0.3.1
pygame==1.9.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
scipy==1.0.0
selenium==3.6.0
setuptools==36.4.0
six==1.11.0
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
Theano==1.0.0
Werkzeug==0.12.2
wheel==0.29.0
wincertstore==0.2
